buttonset: don't mark axes changed on init
This is a leftover from the tablet bits but not needed here.
We don't communicate the initial axis state (maybe we should, though then we'd
likely need a custom axis source for that).

Meanwhile, having all axes marked as changed on init creates a fake axis event
when the Intuos 5 Pad sends a ABS_MISC on the first button press - it triggers
a fake change on the ring.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>