halfkey: Mirrored keyboard accessibility support
Introduce the halfkey accessibility support into libinput.
This allows a touch-typist to access the majority of a keyboard
layout, using just one hand - by utilising the space bar as a
modifier key.

This is convenient, as it is readily accessible to both the right
and left hand using the thumb, whilst also pressing the desired
key.

A space on its own is still achieved by pressing and releasing the bar.

The mirroring comes naturally to touch-typists, and users can rapidly
pick up speed of one handed typing.